MERGE: from f-release8_20040625 Fixes the following: - Reports     - call volume report     - accounts report     - users that are contacts report     - employees report     - ticket report     - Generating multiple reports - Accounts Assets     - Deleting assets - ContactsList popup list (used in various places) - LookupList popup selector - CategoryEditor     - Removing items     - Adding items - Dependancy message when deleting any items - Deleting contacts now remove themselves from action lists - Installing war now properly shows admin user as an employee - Reset buttons removed

+++ src/java/org/aspcfs/modules/accounts/actions/AccountsAssets.java
@@ -281,9 +281,14 @@ public class AccountsAssets extends CFSModule {
       htmlDialog.setTitle("Dark Horse CRM: Account Management - Asset");
       DependencyList dependencies = thisAsset.processDependencies(db);
       htmlDialog.addMessage(dependencies.getHtmlString());
-      htmlDialog.setHeader("The asset you are requesting to delete has the following dependencies within Dark Horse CRM:");
-      htmlDialog.addButton("Delete All", "javascript:window.location.href='AccountsAssets.do?command=Delete&action=delete&orgId=" + orgId + "&id=" + id + "'");
-      htmlDialog.addButton("Cancel", "javascript:parent.window.close()");
+      if (dependencies.canDelete()){
+        htmlDialog.setHeader("The asset you are requesting to delete has the following dependencies within Dark Horse CRM:");
+        htmlDialog.addButton("Delete All", "javascript:window.location.href='AccountsAssets.do?command=Delete&action=delete&orgId=" + orgId + "&id=" + id + "'");
+        htmlDialog.addButton("Cancel", "javascript:parent.window.close()");
+      }else{
+        htmlDialog.setHeader("The asset cannot be deleted because it has the following dependencies within Dark Horse CRM:");
+        htmlDialog.addButton("OK", "javascript:parent.window.close()");
+      }
     } catch (Exception e) {
       errorMessage = e;
     } finally {
